Proyecto con funcionalidad de subida de archivos (Pulir fallos)

+++ ws/ws1/jacaranda_V2/src/main/java/com/jacaranda/repo/PedidoRepository.java
@@ -15,7 +15,7 @@ public interface PedidoRepository extends CrudRepository<Pedido, Integer>{
 			//public  Pedido findPedidoByName(String name);
 
 			//Busca un pedido por id
-			public Pedido findPedidoById(int id);
+			public Pedido findPedidoById(Long id);
 			
 			
 			// Get de todos los pedidos ordenado por nombre
